Raster: zero scratch buffers instead of leaving them uninitialized
